update project structure
-> move data to a seperate repo
-> update the way modules are organized
-> move some code around
-> create a package bhagavad_gita_api
-> fix code quality errors

name of repo = name in PyPI = name of python package = name in DockerHub
