+++ bhagavad_gita_api/__init__.py
@@ -1,0 +1,14 @@
+"""Package bhagavad-gita-api.
+
+Bhagavad Gita API allows any developer to use content from Bhagavad Gita in their applications.
+
+MIT License
+
+Copyright (c) 2021 The Gita Initiative <contact@bhagavadgita.io>
+
+https://github.com/gita/bhagavad-gita-api
+"""
+
+from importlib.metadata import version
+
+__version__ = version(__package__)
