--- src/api/api_v2/api.py
@@ -1,6 +1,0 @@
-from fastapi import APIRouter
-
-from .endpoints import gita
-
-api_router = APIRouter()
-api_router.include_router(gita.router)
