--- src/gita_data/authors.json
@@ -1,65 +1,0 @@
-[
-  {
-    "name": "Swami Ramsukhdas"
-  },
-  {
-    "name": "Swami Chinmayananda"
-  },
-  {
-    "name": "Sri Anandgiri"
-  },
-  {
-    "name": "Sri Dhanpati"
-  },
-  {
-    "name": "Sri Madhavacharya"
-  },
-  {
-    "name": "Sri Neelkanth"
-  },
-  {
-    "name": "Sri Ramanujacharya"
-  },
-  {
-    "name": "Sri Sridhara Swami"
-  },
-  {
-    "name": "Sri Vedantadeshikacharya Venkatanatha"
-  },
-  {
-    "name": "Sri Abhinavgupta"
-  },
-  {
-    "name": "Sri Jayatritha"
-  },
-  {
-    "name": "Sri Madhusudan Saraswati"
-  },
-  {
-    "name": "Sri Purushottamji"
-  },
-  {
-    "name": "Sri Shankaracharya"
-  },
-  {
-    "name": "Sri Vallabhacharya"
-  },
-  {
-    "name": "Swami Tejomayananda"
-  },
-  {
-    "name": "Swami Adidevananda"
-  },
-  {
-    "name": "Swami Gambirananda"
-  },
-  {
-    "name": "Swami Sivananda"
-  },
-  {
-    "name": "Shri Purohit Swami"
-  },
-  {
-    "name": "Dr. S. Sankaranarayan"
-  }
-]
