--- src/gita_data/languages.json
@@ -1,11 +1,0 @@
-[
-    {
-        "language":"english"
-    },
-    {
-        "language":"hindi"
-    },
-    {
-        "language":"sanskrit"
-    }
-]
